requirements.txt
fyers-apiv3>=3.0.0
streamlit>=1.30.0
streamlit-autorefresh>=1.0.1
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
plotly>=5.18.0
python-dotenv>=1.0.0
